Create LMSbyteja.java
An application for Library Management in which read the Student Registration Number and password. Now if the entered credentials are correct then display the name and Number of books issued to that Student or if user does not exist then ask to enter details to create an account. Display the menu to Issue Book / Return Book.
If User wants to get the book issued then ask to enter the Book_name and check in the available book record. If available, then issue the book and update the issued book count and available book record. If the user wants to return the book, then update the available book record and number of books issued to the student. [Note: Create a static list for books in the library and student registration number and password]